(MISC): Moved `RecursiveRustVisitor` into `visitors` package;         Tidying up
