+++ src/main/kotlin/org/rust/lang/core/psi/impl/mixin/RustPathPartImplMixin.kt
@@ -35,28 +35,6 @@ abstract class RustPathPartImplMixin(node: ASTNode) : RustNamedElementImpl(node)
             }
         }
 
-
-    /**
-     *  Returns `true` if this is a fully qualified path.
-     *
-     *  Paths in use items are special, they are implicitly FQ.
-     *
-     *  Example:
-     *
-     *    ```Rust
-     *    use ::foo::bar;   // FQ
-     *    use foo::bar;     // FQ, the same as the above
-     *
-     *    fn main() {
-     *        ::foo::bar;   // FQ
-     *        foo::bar;     // not FQ
-     *    }
-     *    ```
-     *
-     *  Reference:
-     *    https://doc.rust-lang.org/reference.html#paths
-     *    https://doc.rust-lang.org/reference.html#use-declarations
-     */
     override val isFullyQualified: Boolean
         get() {
             val qual = qualifier
@@ -67,19 +45,6 @@ abstract class RustPathPartImplMixin(node: ASTNode) : RustNamedElementImpl(node)
             }
         }
 
-    /**
-     *  Returns true if this path references ancestor module via `self` and `super` chain.
-     *
-     *  Paths can contain any combination of identifiers and self and super keywords.
-     *  However, a path is "well formed" only if it starts with `(self::)? (super::)*`.
-     *  In other words, in `foo::super::bar` the `super` is meaningless and resolves
-     *  to nothing.
-     *
-     *  This check returns true for `(self::)? (super::)*` part of a path.
-     *
-     *  Reference:
-     *    https://doc.rust-lang.org/reference.html#paths
-     */
     override val isModulePrefix: Boolean
         get() {
             val qual = qualifier
